+++ src/main/java/de/gematik/demis/nps/base/profile/package-info.java
@@ -23,6 +23,7 @@ package de.gematik.demis.nps.base.profile;
  *
  * *******
  *
- * For additional notes and disclaimer from gematik and in case of changes by gematik find details in the "Readme" file.
+ * For additional notes and disclaimer from gematik and in case of changes by gematik,
+ * find details in the "Readme" file.
  * #L%
  */